Change submission ipaddress column to BIGINT
Bump plugin schemaVersion to 1.5.1 and change SubmissionRecord.ipaddress from integer to bigInteger. Update Install and existing settings migration to create the column as BIGINT, and add a new migration (m260624_083639_submissions_ipaddress_bigint) to alter the column for existing installs (with a safeDown to revert). DB schema is refreshed after changes.

+++ src/migrations/m251128_041612_settings_1_5.php
@@ -14,7 +14,7 @@ class m251128_041612_settings_1_5 extends Migration {
     public function safeUp(): bool {
       if (!$this->db->tableExists(SubmissionRecord::tableName())){
         $this->createTable(SubmissionRecord::tableName(), [
-          'ipaddress' => $this->integer()->unsigned(),
+          'ipaddress' => $this->bigInteger()->unsigned(),
           'dateCreated' => $this->dateTime()->notNull(),
           'dateUpdated' => $this->dateTime()->notNull(),
           'uid' => $this->uid(),
